refactor graph search
