--- src/agdb/graph_search/path.rs
@@ -1,7 +1,0 @@
-use crate::graph::GraphIndex;
-
-#[derive(Clone)]
-pub(crate) struct Path {
-    pub(crate) elements: Vec<GraphIndex>,
-    pub(crate) cost: u64,
-}
