--- src/agdb/graph_search/path_search_handler.rs
@@ -1,5 +1,0 @@
-use crate::graph::GraphIndex;
-
-pub trait PathSearchHandler {
-    fn process(&self, index: GraphIndex, distance: u64) -> u64;
-}
