--- src/agdb/graph_search/search_control.rs
@@ -1,6 +1,0 @@
-#[allow(dead_code)]
-pub enum SearchControl {
-    Continue(bool),
-    Finish(bool),
-    Stop(bool),
-}
