--- src/agdb/graph_search/search_handler.rs
@@ -1,6 +1,0 @@
-use super::search_control::SearchControl;
-use crate::graph::GraphIndex;
-
-pub trait SearchHandler {
-    fn process(&self, index: &GraphIndex, distance: &u64) -> SearchControl;
-}
