+++ src/agdb/graph_search/search_impl.rs
@@ -1,13 +1,24 @@
-use super::search_control::SearchControl;
-use super::search_handler::SearchHandler;
-use super::search_index::SearchIndex;
-use super::search_iterator::SearchIterator;
+use super::SearchControl;
+use super::SearchHandler;
 use crate::collections::bit_set::BitSet;
 use crate::graph::GraphData;
 use crate::graph::GraphImpl;
 use crate::graph::GraphIndex;
 use std::marker::PhantomData;
 use std::mem::swap;
+
+#[derive(Clone, Copy)]
+pub(crate) struct SearchIndex {
+    pub(crate) index: GraphIndex,
+    pub(crate) distance: u64,
+}
+
+pub(crate) trait SearchIterator {
+    fn expand_edge<Data: GraphData>(index: GraphIndex, graph: &GraphImpl<Data>) -> GraphIndex;
+    fn expand_node<Data: GraphData>(index: GraphIndex, graph: &GraphImpl<Data>) -> Vec<GraphIndex>;
+    fn new(stack: &mut Vec<SearchIndex>) -> Self;
+    fn next(&mut self) -> Option<SearchIndex>;
+}
 
 pub(crate) struct SearchImpl<'a, Data, SearchIt>
 where
@@ -133,5 +144,35 @@ where
         self.visited.insert(index.index.as_u64());
 
         visited
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    #[allow(clippy::clone_on_copy)]
+    fn derived_from_clone() {
+        let index = SearchIndex {
+            index: GraphIndex(1),
+            distance: 10,
+        };
+        let other = index.clone();
+
+        assert_eq!(index.index, other.index);
+        assert_eq!(index.distance, other.distance);
+    }
+
+    #[test]
+    fn derived_from_copy() {
+        let index = &SearchIndex {
+            index: GraphIndex(1),
+            distance: 10,
+        };
+        let other = *index;
+
+        assert_eq!(index.index, other.index);
+        assert_eq!(index.distance, other.distance);
     }
 }
